requirements.txt
python-can>=4.6.1
cantools>=40.7.1
rich>=13.0.0